Refactored files to match renamed component names
